api_test_work/serializers.py
from rest_framework import serializers

from .models import HackerNews


class HackerNewsListSerializer(serializers.ModelSerializer):
    """Вывод списка актеров и режиссеров"""
    class Meta:
        model = HackerNews
        fields = ('__all__')