requirements.txt
amqp==2.6.0
asgiref==3.2.10
asposestorage==1.0.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-filter==2.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
future==0.18.2
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
kombu==4.6.11
lxml==4.5.2
Markdown==3.2.2
multidict==4.7.6
Pillow==7.2.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
uWSGI==2.0.19.1
vine==1.3.0
yarl==1.4.2
zope.interface==5.1.0
